mixed_libft

+++ corewar/mixed_libft/ft_printf/srcs/ftpf_ch_conv.c
@@ -6,7 +6,7 @@
 /*   By: lcharbon <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/19 15:09:25 by lcharbon          #+#    #+#             */
-/*   Updated: 2017/06/08 02:36:47 by lcharbon         ###   ########.fr       */
+/*   Updated: 2017/07/24 00:49:23 by lcharbon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ static void			ftpf_width(t_data *data, char **str)
 {
 	char			*tmp;
 
-	if (data->data[1] > ft_strlen(*str))
+	if (data->data[1] > (int)ft_strlen(*str))
 	{
 		if (data->data[3] == 0 || data->data[4] == 1)
 			tmp = ft_strnewch(data->data[1] - 1, ' ');
@@ -39,16 +39,16 @@ int					ftpf_ch_conv(t_data *data, va_list ap)
 	ft_memset(tmp, '\0', 5);
 	ftpf_valist_char(data, ap, tmp);
 	str = ft_strdup(tmp);
-	if (ft_strlen(tmp) == 0)
+	if ((int)ft_strlen(tmp) == 0)
 		a++;
-	if (ft_strlen(tmp) == 0 && data->data[4] == 1)
+	if ((int)ft_strlen(tmp) == 0 && data->data[4] == 1)
 		ft_putchar_fd(0, data->fd);
 	if (data->data[1] > 0)
 		ftpf_width(data, &str);
 	ft_putstr_fd(str, data->fd);
-	if (ft_strlen(tmp) == 0 && data->data[4] == 0)
+	if ((int)ft_strlen(tmp) == 0 && data->data[4] == 0)
 		ft_putchar_fd(0, data->fd);
-	a += ft_strlen(str);
+	a += (int)ft_strlen(str);
 	free(str);
 	return (a);
 }
