+++ corewar/mixed_libft/ft_printf/srcs/ftpf_expo_conv.c
@@ -6,7 +6,7 @@
 /*   By: lcharbon <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/07 01:34:15 by lcharbon          #+#    #+#             */
-/*   Updated: 2017/06/10 01:36:33 by lcharbon         ###   ########.fr       */
+/*   Updated: 2017/07/24 00:53:42 by lcharbon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,15 +18,15 @@ static void			ftpf_width_noless(t_data *data, intmax_t arg, char **str)
 	char			*tmp2;
 
 	if (data->data[3] == 0 || data->data[0] >= 0)
-		tmp = ft_strnewch(data->data[1] - ft_strlen(*str), ' ');
+		tmp = ft_strnewch(data->data[1] - (int)ft_strlen(*str), ' ');
 	else if (arg >= 0 && *str[0] != ' ' && *str[0] != '+')
-		tmp = ft_strnewch(data->data[1] - ft_strlen(*str), '0');
+		tmp = ft_strnewch(data->data[1] - (int)ft_strlen(*str), '0');
 	else
 	{
-		tmp = ft_strnewch(data->data[1] - ft_strlen(*str) + 1, '0');
+		tmp = ft_strnewch(data->data[1] - (int)ft_strlen(*str) + 1, '0');
 		tmp[0] = *str[0];
 		tmp2 = *str;
-		*str = ft_strsub(*str, 1, ft_strlen(*str) - 1);
+		*str = ft_strsub(*str, 1, (int)ft_strlen(*str) - 1);
 		ft_strdel(&tmp2);
 	}
 	*str = ft_strjoindel(tmp, *str, 3);
@@ -36,7 +36,7 @@ static void			ftpf_width_less(t_data *data, char **str)
 {
 	char			*tmp;
 
-	if (!(tmp = ft_strnewch(data->data[1] - ft_strlen(*str), ' ')))
+	if (!(tmp = ft_strnewch(data->data[1] - (int)ft_strlen(*str), ' ')))
 		return ;
 	*str = ft_strjoindel(*str, tmp, 3);
 }
@@ -84,13 +84,13 @@ int					ftpf_expo_conv(t_data *data, va_list ap)
 	ftpf_expo_pre(&str, arg, data->data[0]);
 	if (data->data[5] == 1 || data->data[6] == 1)
 		ftpf_flag(data, arg, &str);
-	if (data->data[1] > ft_strlen(str) && data->data[4] == 1)
+	if (data->data[1] > (int)ft_strlen(str) && data->data[4] == 1)
 		ftpf_width_less(data, &str);
-	if (data->data[1] > ft_strlen(str) && data->data[4] == 0)
+	if (data->data[1] > (int)ft_strlen(str) && data->data[4] == 0)
 		ftpf_width_noless(data, arg, &str);
 	ft_maj_e(str, data->conv);
 	ft_putstr_fd(str, data->fd);
-	ret = ft_strlen(str);
+	ret = (int)ft_strlen(str);
 	free(str);
 	return (ret);
 }
